third_party/WebKit/LayoutTests/inspector-protocol/css/css-coverage-poll-expected.txt

Class Selector

ID Selector

 =>
Coverage delta (5 entries)
    css/resources/coverage.css: +0-35 +36-56 +57-74 +141-191 +327-517
loadStylesheet('resources/coverage2.css') =>
Coverage delta (1 entries)
    css/resources/coverage2.css: +95-151
useMoreCSS('usedSomewhatLater') =>
Coverage delta (1 entries)
    css/resources/coverage2.css: +34-93
Coverage delta (1 entries)
    css/resources/coverage2.css: +153-209
